Rename project
- Rename project
- Rename namespace
- Rename main class
- Edit url connection construction
- All words translate to English
