+++ Notificator/App.config
@@ -4,8 +4,8 @@
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
   </startup>
   <appSettings>
-    <add key="username" value="agajanh" />
-    <add key="url" value="http://asmanoky.com/notification.php" />
+    <add key="username" value="evgeniy" />
+    <add key="url" value="http://127.0.0.1:8181/api/notification/" />
     <add key="interval" value="30" />
     <add key="ClientSettingsProvider.ServiceUri" value="" />
   </appSettings>
