+++ Notificator/Notificator.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{3E59353B-7478-43EB-A077-241BB0D7D9E6}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>Notifier</RootNamespace>
-    <AssemblyName>Notifier</AssemblyName>
+    <RootNamespace>Notificator</RootNamespace>
+    <AssemblyName>Notificator</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -68,17 +68,17 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Data.cs" />
-    <Compile Include="Notifier.cs">
+    <Compile Include="Notificator.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Notifier.Designer.cs">
-      <DependentUpon>Notifier.cs</DependentUpon>
+    <Compile Include="Notificator.Designer.cs">
+      <DependentUpon>Notificator.cs</DependentUpon>
     </Compile>
     <Compile Include="Notification.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Notifier.resx">
-      <DependentUpon>Notifier.cs</DependentUpon>
+    <EmbeddedResource Include="Notificator.resx">
+      <DependentUpon>Notificator.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
