+++ Notificator/Properties/Resources.resx
