+++ Notificator/Properties/Settings.settings
