+++ README.md
@@ -1,1 +1,1 @@
-Notifier
+Notificator
